Restructured Projects, added titles, and updated new structure
