+++ src/features/common/Layout/UserLayout/UserLayout.module.scss
@@ -4,11 +4,8 @@
   position: relative;
   display: flex;
 }
-.profilePage {
+.profilePageWrapper {
   width: 100%;
-  max-width: 1000px;
-  margin-top: 80px;
-  padding: 40px;
   min-height: 80vh;
 }
 
@@ -83,4 +80,10 @@
       background-color: $blueishGrey;
     }
   }
+}
+
+.widgetIFrame {
+  min-height: 100vh;
+  width: 100%;
+  max-width: calc(100vw - 256px);
 }
