+++ src/features/user/Account/components/AccountRecord.tsx
@@ -1,9 +1,9 @@
 import React, { ReactElement } from 'react';
-import styles from '../../styles/AccountHistory.module.scss';
-import getFormatedCurrency from '../../../../../utils/countryCurrency/getFormattedCurrency';
-import formatDate from '../../../../../utils/countryCurrency/getFormattedDate';
-import { getFormattedNumber } from '../../../../../utils/getFormattedNumber';
-import i18next from '../../../../../../i18n';
+import styles from '../AccountHistory.module.scss';
+import getFormatedCurrency from '../../../../utils/countryCurrency/getFormattedCurrency';
+import formatDate from '../../../../utils/countryCurrency/getFormattedDate';
+import { getFormattedNumber } from '../../../../utils/getFormattedNumber';
+import i18next from '../../../../../i18n';
 
 const { useTranslation } = i18next;
