+++ src/features/user/CompleteSignup/index.tsx
@@ -121,7 +121,7 @@ export default function CompleteSignup() {
         console.log('in 401-> unauthenticated user / invalid token');
         setUser(false);
         setSubmit(false);
-        logoutUser();
+        logoutUser(`${process.env.NEXTAUTH_URL}/`);
         loginWithRedirect({
           redirectUri: `${process.env.NEXTAUTH_URL}/login`,
           ui_locales: localStorage.getItem('language') || 'en',
@@ -195,7 +195,7 @@ export default function CompleteSignup() {
         <div className={requestSent ? styles.signupRequestSent : styles.signup}>
           {/* header */}
           <div className={styles.header}>
-            <div onClick={() => logoutUser()} className={styles.headerBackIcon}>
+            <div onClick={() => logoutUser(`${process.env.NEXTAUTH_URL}/`)} className={styles.headerBackIcon}>
               <CancelIcon color={styles.primaryFontColor} />
             </div>
             <div className={styles.headerTitle}>
