+++ src/features/user/Profile/styles/Profile.module.scss
@@ -1,10 +1,5 @@
 @import './../../../../theme/theme';
 
-.widgetIFrame {
-  min-height: 100vh;
-  width: 100%;
-  max-width: calc(100vw - 256px);
-}
 .settingsIcon {
   display: flex;
   position: absolute;
