+++ src/features/user/TreeMapper/TreeMapper.module.scss
@@ -1,5 +1,4 @@
-@import './../../../../theme/theme';
-
+@import '../../../theme/theme';
 .pageContainer {
   display: flex;
   flex-direction: column;
