+++ src/features/user/Widget/EmbedModal.module.scss
@@ -1,4 +1,4 @@
-@import './../../../../theme/theme';
+@import './../../../theme/theme';
 
 .modal {
   //   top: 80px;
